Fix Rubocop Linters.
