+++ interfaces/books.rb
@@ -14,7 +14,6 @@ class Books
   def find_by_id(id)
     @books.find { |book| book.id == id }
   end
-  
 
   def create_book
     puts 'Please, enter book information below:'
